Reorder Dockerfile to better utilize layer caching and minimize final image size (#206)
* Add support for debian tag variants with default tag set to buster-slim

* Move locale generation to beginning of Dockerfile

As locale-gen takes a lot of time when creating a lot of locales
it is best to put it as early as possible in the Dockerfile to
utilize the layer caching mechanism.

* Consolidate package operations into single layers

It is necessary to do the package cache cleanup in the same command
as package installation to not waste image space. Cleaning the apt
cache later in the Dockerfile does not decrease the total image size.

* Remove unnecessary apt cache cleanup

* Copy all scripts in one go and consolidate RUN commands

Best practice is to use COPY to copy files into the image. ADD
is ambiguous as it can extract archives.

* Make the postgresql data dir a volume

* Whitespace

* Add ping for use in replication setup